% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolverine_habitat}
\alias{wolverine_habitat}
\title{Wolverine Habitat Raster}
\format{
\subsection{\code{wolverine_habitat}}{

A RasterLayer with 82269 pixels:
\describe{
\item{values}{Probability of wolverine use}
}
}
}
\source{
Raster comes from original rSPACE package and source data files that
were provided as an example for package use
}
\usage{
wolverine_habitat
}
\description{
A raster showing the habitat use for wolverines in the bitterroot mountain
range. This habitat use layer is representative of persistant spring snow
cover as has been shown to predict wolverine occurrance.
}
\keyword{datasets}
